Updated trigger fragment metadata to fix integer definition mismatch

+++ sbndaq-artdaq-core/Overlays/ICARUS/ICARUSTriggerUDPFragment.hh
@@ -22,7 +22,7 @@ class icarus::ICARUSTriggerUDPFragmentMetadata
 
 public:  
   ICARUSTriggerUDPFragmentMetadata() {}
-  ICARUSTriggerUDPFragmentMetadata(int n, int ev, int sec, long nsec,int wr_n, int wr_ev, long wr_sec, long wr_nsec) : name(n), event_no(ev), seconds(sec), nanoseconds(nsec), wr_name(wr_n), wr_event_no(wr_ev), wr_seconds(wr_sec), wr_nanoseconds(wr_nsec) {}
+  ICARUSTriggerUDPFragmentMetadata(int n, int ev, int sec, long nsec,int wr_n, int wr_ev, long long wr_sec, long long wr_nsec) : name(n), event_no(ev), seconds(sec), nanoseconds(nsec), wr_name(wr_n), wr_event_no(wr_ev), wr_seconds(wr_sec), wr_nanoseconds(wr_nsec) {}
 
   int getName() const
   { return name; }
@@ -42,7 +42,7 @@ public:
   int getWREventNo() const
   { return wr_event_no; }
 
-  int getWRSeconds() const
+  long getWRSeconds() const
   { return wr_seconds; }
 
   long getWRNanoSeconds() const
@@ -93,7 +93,7 @@ public:
   int getWREventNo() const
   { return Metadata()->getWREventNo(); }
 
-  int getWRSeconds() const
+  long getWRSeconds() const
   { return Metadata()->getWRSeconds(); }
 
   long getWRNanoSeconds() const
